face-verification/requirements.txt
streamlit==1.47.1
deepface==0.0.93
opencv-python-headless==4.12.0.88
tensorflow==2.19.0
numpy==2.1.3
Pillow
keras==3.10.0
tf-keras==2.19.0